+++ CHANGELOG.md
@@ -20,6 +20,7 @@
  * Support `ignore` property in `Column` annotation
  * Refactor `SqlProvider.paginate` method
  * Support left join for select query
+ * Support right join and inner join for select query
 
 ## 1.0.3